Move FeatureInEra to Cardano.Api.Eras module
