+++ cardano-api/internal/Cardano/Api/Feature.hs
@@ -6,8 +6,6 @@
 
 module Cardano.Api.Feature
   ( FeatureValue (..)
-  , FeatureInEra(..)
-  , featureInShelleyBasedEra
   , valueOrDefault
   , asFeatureValue
   , asFeatureValueInShelleyBasedEra
@@ -15,38 +13,6 @@ module Cardano.Api.Feature
   ) where
 
 import           Cardano.Api.Eras
-
-import           Data.Kind
-
--- | A class for features that are supported in some eras but not others.
-class FeatureInEra (feature :: Type -> Type) where
-  -- | Determine the value to use for a feature in a given 'CardanoEra'.
-  -- Note that the negative case is the first argument, and the positive case is the second as per
-  -- the 'either' function convention.
-  featureInEra :: ()
-    => a                    -- ^ Value to use if the feature is not supported in the era
-    -> (feature era -> a)   -- ^ Function to get thealue to use if the feature is supported in the era
-    -> CardanoEra era       -- ^ Era to check
-    -> a                    -- ^ The value to use
-
-instance FeatureInEra ShelleyBasedEra where
-  featureInEra no yes = \case
-    ByronEra    -> no
-    ShelleyEra  -> yes ShelleyBasedEraShelley
-    AllegraEra  -> yes ShelleyBasedEraAllegra
-    MaryEra     -> yes ShelleyBasedEraMary
-    AlonzoEra   -> yes ShelleyBasedEraAlonzo
-    BabbageEra  -> yes ShelleyBasedEraBabbage
-    ConwayEra   -> yes ShelleyBasedEraConway
-
--- | Determine the value to use for a feature in a given 'ShelleyBasedEra'.
-featureInShelleyBasedEra :: ()
-  => FeatureInEra feature
-  => a
-  -> (feature era -> a)
-  -> ShelleyBasedEra era
-  -> a
-featureInShelleyBasedEra no yes = featureInEra no yes . shelleyBasedToCardanoEra
 
 -- | A value of type @'FeatureValue' feature era a@ is either:
 data FeatureValue feature era a where
